+++ workspaces/servicenow/plugins/servicenow-backend/README.md
@@ -1,6 +1,9 @@
-# servicenow
+# ServiceNow Backend Plugin
 
-This plugin backend was templated using the Backstage CLI. You should replace this text with a description of your plugin backend.
+This is the **backend implementation** of the ServiceNow plugin for Backstage.  
+It provides an API that securely connects to your ServiceNow instance and returns incident data for catalog entities.
+
+---
 
 ## Installation
 
@@ -25,4 +28,6 @@ This plugin backend can be started in a standalone mode from directly in this
 package with `yarn start`. It is a limited setup that is most convenient when
 developing the plugin backend itself.
 
-If you want to run the entire project, including the frontend, run `yarn start` from the root directory.
+## Configuration
+
+Refer to [Configuration.md](../../docs/Configuration.md) for detailed setup instructions, including authentication and ServiceNow instance configuration.
